+++ packages/coinstac-pipeline/src/pipeline-manager-helpers.js
